### 2023-02-11 #5146 @editorialbot
[PRE REVIEW]: CoReCon: an open, community-powered collection of Reionization constraints
**Submitting author:** @EGaraldi (<a href="http://orcid.org/0000-0002-6021-7020">Enrico Garaldi</a>)
**Repository:** https://github.com/EGaraldi/corecon
**Branch with paper.md** (empty if default branch):
**Version:** 0.6.1
