### 2023-02-11 #5146 @dfm
@EGaraldi — Thanks for your submission! As the managing editor for this "track" of JOSS submissions, I'm taking a first pass through the code and paper, and I'm working on recruiting an editor to handle your submission. I'll keep you posted on progress!

Since this has previously made it through [most of a review](https://github.com/openjournals/joss-reviews/issues/2495), could you briefly summarize the changes since that review here in this thread? Thanks!
