+++ corecon/data/UV_slope/Endsley_et_al_2022.py
@@ -1,0 +1,121 @@
+dictionary_tag         = "Endsley et al. 2022 (subm.)"
+
+reference              = "Endsley, Stark, Whitler, Topping, Chen, Plat, Chisholm, Charlot, arXiv (2022)"
+
+url                    = "https://arxiv.org/pdf/2208.14999.pdf"
+
+description            = \
+"""
+sed on CEERS JWST/NIRCam data of 118 UV-faint (median MUV=−19.6) z∼6.5−8 galaxies
+"""
+
+data_structure         = "points" #grid or points
+
+extracted              = False
+
+ndim                   = 1
+
+dimensions_descriptors = ["redshift"]
+
+axes                   = [10.11,  8.95, 10.71,  9.16,  8.28,  8.92,  9.  ,  6.7 ,  7.2 ,
+                          7.3 ,  7.3 ,  7.1 ,  6.6 ,  6.8 ,  7.1 ,  6.8 ,  6.6 ,  7.8 ,
+                          6.8 ,  6.5 ,  6.7 ,  7.3 ,  7.4 ,  7.4 ,  6.4 ,  6.4 ,  7.5 ,
+                          7.  ,  6.8 ,  7.  ,  6.7 ,  6.5 ,  6.9 ,  7.3 ,  7.8 ,  6.6 ,
+                          6.3 ,  6.6 ,  7.7 ,  7.1 ,  6.7 ,  7.6 ,  6.4 ,  6.5 ,  6.3 ,
+                          6.9 ,  7.  ,  6.6 ,  7.8 ,  6.6 ,  7.2 ,  7.  ,  6.8 ,  6.7 ,
+                          7.7 ,  6.7 ,  7.  ,  6.4 ,  6.4 ,  6.7 ,  7.9 ,  6.6 ,  6.6 ,
+                          7.4 ,  7.1 ,  6.5 ,  7.9 ,  7.6 ,  6.8 ,  6.8 ,  6.8 ,  6.6 ,
+                          7.8 ,  6.7 ,  6.8 ,  7.4 ,  6.6 ,  6.5 ,  6.7 ,  6.5 ,  6.7 ,
+                          7.2 ,  6.7 ,  7.4 ,  7.2 ,  6.6 ,  6.7 ,  6.7 ,  6.8 ,  6.9 ,
+                          7.9 ,  6.5 ,  6.9 ,  6.7 ,  7.9 ,  7.3 ,  7.4 ,  6.8 ,  6.5 ,
+                          6.8 ,  6.8 ,  7.4 ,  7.6 ,  6.7 ,  6.4 ,  7.  ,  6.5 ,  6.6 ,
+                          6.4 ,  6.3 ,  7.4 ,  7.2 ,  6.8 ,  6.8 ,  7.1 ,  7.3 ,  7.1 ,
+                          6.7 ,  8.  ,  6.9 ,  6.3 ,  7.1 ,  7.8 ]
+
+values                 = [-2.2 , -1.9 , -1.8 , -1.8 , -1.7 , -2.2 , -2.4 , -2.94, -3.1 ,
+                          -1.83, -2.01, -2.25, -2.21, -1.75, -1.48, -2.2 , -2.67, -2.19,
+                          -1.41, -1.54, -2.09, -2.3 , -2.4 , -2.4 , -1.96, -1.91, -2.77,
+                          -2.13, -0.68, -0.87, -1.44, -1.91, -1.11, -2.38, -1.8 , -2.21,
+                          -2.05, -2.86, -2.03, -1.37, -2.55, -1.94, -2.1 , -1.98, -2.24,
[…]
