### 2023-03-20 #5146 @EGaraldi
> @EGaraldi — I wanted to check in to see if you could please summarize what has changed since the last review to help us recruit reviewers?

Sorry, I missed this question among the other thread posts. Essentially, some personal issues prevented me from replying properly to the referees' comments of my last submission, which was therefore closed. Now that this is resolved, I replied and implemented the referees' comments, as well as implemented the following changes:

- improved, up-to-date collection of constraints provided
- new additional user interface for the retrieval of properties, that are now generated automatically with self-explanatory names
- easier and automated update system, which allows to update the content provided without having to update the python package itself (unless the code itself is changed)
- new available constraints provided
- new user-facing and dev-facing utility functions to retrieve, filter and convert data
- easier data retrieval through synonyms, i.e. different strings pointing to the same dataset
- improved documentation
