### 2023-03-31 cbfea64
bugfix in filters for specific data shape
