### 2023-04-24 #5146 @eloisabentivegna
Dear @EGaraldi, thanks for providing the background for the old submission, and your updates. It's nice that you've already taken care of the issues brought up by @iliantiliev and @afialkov.

Even though the reviews for the last submission of this package are still available at https://github.com/openjournals/joss-reviews/issues/2495, I will start a new review to recap, discuss the updates, and give the reviewers a chance to examine and respond to your changes.
