### 2023-04-24 #5407 @eloisabentivegna
@iliantiliev, @afialkov, you've been added to this issue as you were reviewers for submission https://github.com/openjournals/joss-reviews/issues/2495, and the author @egaraldi has resubmitted a new version of CoReCon for publication, which addresses the questions you raised under the old submission.

It would be of great help if you could take a look at the updates and let me know if you have further questions or can proceed with your recommendation. Some details have already been discussed in [the pre-review issue](https://github.com/openjournals/joss-reviews/issues/5146).
