### 2023-05-20 #5407 @eloisabentivegna
@iliantiliev, @afialkov: could you let me know your position regarding the new version of CoReCon? A quick check that the changes you have requested were implemented, along with your publication recommendations, are all that is needed.
