### 2023-05-23 #5407 @Stefan-Heimersheim
Hi @eloisabentivegna: @afialkov and I checked the new version and saw that everything is implemented, the requested changes in code ([my issues here](https://github.com/EGaraldi/corecon/issues?q=is%3Aissue+author%3AStefan-Heimersheim)) and the paper ([Anastasia's issue here](https://github.com/EGaraldi/corecon/issues/3)).

Would you like to add me as reviewer again as in the [original review](https://github.com/openjournals/joss-reviews/issues/2495)?
