+++ paper/paper.bib
@@ -13,7 +13,8 @@ archivePrefix = {arXiv},
        eprint = {2005.06505},
  primaryClass = {astro-ph.IM},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2020arXiv200506505P},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+      doi = {10.21105/joss.02308}
 }
 
 
