### 2023-06-02 59fba50
fix to filter_by_redshift_range for numpy>1.21
