### 2023-06-02 #5407 @EGaraldi
@eloisabentivegna I have fixed an additional minor issue pointed out by @Stefan-Heimersheim ([here](https://github.com/EGaraldi/corecon/issues/5)) and accepted your pull request.
