### 2023-06-15 a2576ab
Carry out a few typographic and stylistic tweaks.
