+++ paper/paper.md
@@ -20,9 +20,9 @@ bibliography: paper.bib
 
 # Summary
 
-The number of available constraints on the Universe during and before cosmic reionization is grapidly growing. 
+The number of available constraints on the Universe during and before cosmic reionization is rapidly growing. 
 These are often scattered across inhomogeneous formats, unit systems and sampling strategies. In this paper I introduce 
-`CoReCon`, a python package designed to provide a growing set of constraints on key physical quantities related to the EoR 
+`CoReCon`, a python package designed to provide a growing set of constraints on key physical quantities related to the Epoch of Reionization 
 and a platform for the high-redshift research community to collect and store, in an open way, current and forthcoming 
 observational constraints. 
 
@@ -52,7 +52,7 @@ or unwilling to share the data, a tedious and error-prone task and a sub-standar
 
 I tackle these
 issues through a systematic collection of published constraints on physical properties of the Universe during 
-the EoR. This collection is wrapped in a python module named `CoReCon (acronym for Collection of Reionization Constraint) that 
+the EoR. This collection is wrapped in a python module named `CoReCon` (acronym for Collection of Reionization Constraint) that 
 I present in this paper. I start by introducing the goals and design choices of `CoReCon`, then move 
 on to a description of its features, followed by a review of the available constraints at the time of 
 publication. I close this paper discussing desirable future developments and with a pledge to the community.
@@ -67,7 +67,7 @@ data. Second, and foremost, `CoReCon` provides a platform for the high-redshift
 research community to collect and store, in an open way, such observational constraints. I do 
 so by providing a Python
 infrastructure, which is able to load formatted data files and provides simple utility functions
-to deal with such data. The data files loaded by `CoReCon` are purposedly simple in their form and 
+to deal with such data. The data files loaded by `CoReCon` are purposely simple in their form and 
 as complete as possible in their content, in order to collect all the relevant information in one place. 
 Notably, they are required to contain a URL to the original publication and a short description of 
 the methods used to retrieve the constraints from observed data. They are also allowed to contain any 
@@ -77,14 +77,14 @@ Ideally, once a new constraint is published, the author of the publication will 
 with the relevant data. If this procedure becomes customary, `CoReCon` will serve as an up-to-date 
 repository of easy-to-retrieve constraints.
 
-To our knowledge, this is the first module of its kind - at least in the EoR community. With a similar 
+To our knowledge, this is the first module of its kind -- at least in the EoR community. With a similar 
 spirit, there exist a collection of all the known quasars above a redshift of 5.7 (@qsolist) and a 
-compilation of galaxy data for the specific purpose of validating the VELOCIraptor halo finder (@velociraptor_validation_data). Additionally, 
+compilation of galaxy data for the specific purpose of validating the `VELOCIraptor` halo finder (@velociraptor_validation_data). Additionally, 
 an effort toward openness of research in the EoR field recently materialized into an open analysis
 pipeline for the reduction of spectra taken in most of the major telescopes in the world (@pypelt).
 
 In its development version, `CoReCon`  has been used for the scientific analysis of the THESAN simulations 
-(@Garaldi2022, @Kannan2022, @Smith2022) and is being used in upcoming scietific projects. 
+(@Garaldi2022, @Kannan2022, @Smith2022) and is being used in upcoming scientific projects. 
 
 
 # Features
@@ -109,7 +109,7 @@ errors[^1]), and a short description of the constraints themselves and of the me
 
 The `CoReCon` module can be easily installed via `pip` and is fully documented online at [https://corecon.readthedocs.io/en/latest/](https://corecon.readthedocs.io/en/latest/). 
 `CoReCon` autonomously fetches updates to the constraints at startup (but limited to once every 24 hours or when manually 
[…]
