### 2023-06-22 #5407 @EGaraldi
I have done the steps listed above.
The version tag is 1.0.
The DOI is 10.5281/zenodo.8069253 (https://doi.org/10.5281/zenodo.8069253)
