### 2023-06-30 f762a65
Expanding journal name macros
The JOSS TeX template doesn't include the ADS journal name macros, so we need to expand them.
