+++ paper/paper.bib
@@ -255,7 +255,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Harikane22a,
        author = {{Harikane}, Yuichi and {Inoue}, Akio K. and {Mawatari}, Ken and {Hashimoto}, Takuya and {Yamanaka}, Satoshi and {Fudamoto}, Yoshinobu and {Matsuo}, Hiroshi and {Tamura}, Yoichi and {Dayal}, Pratika and {Yung}, L.~Y. Aaron and {Hutter}, Anne and {Pacucci}, Fabio and {Sugahara}, Yuma and {Koekemoer}, Anton M.},
         title = "{A Search for H-Dropout Lyman Break Galaxies at z 12-16}",
-      journal = {\apj},
+      journal = {The Astrophysical Journal},
      keywords = {Galaxy evolution, Galaxy formation, High-redshift galaxies, 594, 595, 734, Astrophysics - Astrophysics of Galaxies},
          year = 2022,
         month = apr,
@@ -356,7 +356,7 @@ archivePrefix = {arXiv},
 @ARTICLE{RB2022,
        author = {{Roberts-Borsani}, Guido and {Morishita}, Takahiro and {Treu}, Tommaso and {Leethochawalit}, Nicha and {Trenti}, Michele},
         title = "{The Physical Properties of Luminous z {\ensuremath{\gtrsim}} 8 Galaxies and Implications for the Cosmic Star Formation Rate Density from 0.35 deg$^{2}$ of (Pure-)Parallel HST Observations}",
-      journal = {\apj},
+      journal = {The Astrophysical Journal},
      keywords = {Early universe, High-redshift galaxies, Reionization, Star formation, 435, 734, 1383, 1569, Astrophysics - Astrophysics of Galaxies},
          year = 2022,
         month = mar,
@@ -376,7 +376,7 @@ archivePrefix = {arXiv},
 @article{Kannan2022,
   author        = {{Kannan}, R. and {Garaldi}, E. and {Smith}, A. and {Pakmor}, R. and {Springel}, V. and {Vogelsberger}, M. and {Hernquist}, L.},
   title         = {{Introducing the {\textsc{thesan}} project: radiation-magnetohydrodynamic simulations of the epoch of reionization}},
-  journal       = {\mnras},
+  journal       = {Monthly Notices of the Royal Astronomical Society},
   keywords      = {radiative transfer, methods: numerical, galaxies: high-redshift, dark ages, reionization, first stars, Astrophysics - Astrophysics of Galaxies, Astrophysics - Cosmology and Nongalactic Astrophysics},
   year          = 2022,
   month         = apr,
@@ -394,7 +394,7 @@ archivePrefix = {arXiv},
 @article{Garaldi2022,
   author        = {{Garaldi}, E. and {Kannan}, R. and {Smith}, A. and {Springel}, V. and {Pakmor}, R. and {Vogelsberger}, M. and {Hernquist}, L.},
   title         = {{The {\textsc{thesan}} project: properties of the intergalactic medium and its connection to reionization-era galaxies}},
-  journal       = {\mnras},
+  journal       = {Monthly Notices of the Royal Astronomical Society},
   keywords      = {radiative transfer, methods: numerical, galaxies: high-redshift, dark ages, reionization, first stars, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Astrophysics of Galaxies},
   year          = 2022,
   month         = jun,
@@ -412,7 +412,7 @@ archivePrefix = {arXiv},
 @article{Smith2022,
   author        = {{Smith}, A. and {Kannan}, R. and {Garaldi}, E. and {Vogelsberger}, M. and {Pakmor}, R. and {Springel}, V. and {Hernquist}, L.},
[…]
