### 2023-06-30 840d7da
Fixing some typos and syntax errors
