+++ paper/paper.bib
@@ -84,7 +84,7 @@ archivePrefix = {arXiv},
 }
 
 @software{pandas,
-    author       = {The pandas development team,},
+    author       = {The pandas development team},
     title        = {pandas-dev/pandas: Pandas},
     month        = feb,
     year         = 2020,
