+++ corecon/data/mass_gas_metallicity_relation/Bunker_et_al_2023.py
@@ -1,0 +1,34 @@
+dictionary_tag         = "Bunker et al. 2023"
+
+reference              = "Bunker, Saxena, Cameron, Willott, Curtis-Lake, et al.;al.; MNRAS 518, 425 (2023)"
+
+url                    = "https://ui.adsabs.harvard.edu/abs/2023arXiv230207256B/abstract"
+
+description            = \
+"""
+Based on JADES JWST/NIRSpec spectroscopy of GN-z11 (M_UV=-21.5, z=10.603). Note: the metallicity has been converted to 12+log(O/H) assuming Z_sun = 0.0127
+"""
+
+data_structure         = "points" #grid or points
+
+extracted              = False
+
+ndim                   = 2
+
+dimensions_descriptors = ["redshift", "log stellar mass [Msun]"]
+
+axes                   = [[10.603, 8.73]]
+
+err_left               = [[0.0, 0.06]]
+
+err_right              = [[0.0, 0.06]]
+
+values                 = [7.65]
+
+err_up                 = [0.08]
+
+err_down               = [0.07]
+
+upper_lim              = False
+
+lower_lim              = False
