+++ corecon/data/mass_gas_metallicity_relation/Trump_et_al_2023.py
@@ -1,0 +1,34 @@
+dictionary_tag         = "Trump et al. 2023"
+
+reference              = "Trump, Arrabal Haro, Simons, Backhaus, Amorin, Dickinson, et al.; ApJ 945, 35 (2023)"
+
+url                    = "https://ui.adsabs.harvard.edu/abs/2023ApJ...945...35T/abstract"
+
+description            = \
+"""
+From emission-line flux ratio measurements for five z > 5 galaxies in the JWST/NIRSpec Early Release Observationsof SMACS 0723. Stellar masses from Carnall et al. 2022.
+"""
+
+data_structure         = "points" #grid or points
+
+extracted              = False
+
+ndim                   = 2
+
+dimensions_descriptors = ["redshift", "log stellar mass [Msun]"]
+
+axes                   = [[8.4957, 7.10], [6.3792, 7.39], [7.6651, 8.23], [7.6597, 7.72]]
+
+err_left               = [[0.0, 0.12], [0.0, 0.03], [0.0, 0.09], [0.0, 0.05]]
+
+err_right              = [[0.0, 0.14], [0.0, 0.04], [0.0, 0.08], [0.0, 0.06]]
+
+values                 = [7.49, 7.87, 8.10, 7.75]
+
+err_up                 = [0.26, 0.17, 0.17, 0.19]
+
+err_down               = [0.26, 0.17, 0.17, 0.19]
+
+upper_lim              = False
+
+lower_lim              = False
