+++ corecon/data/mass_gas_metallicity_relation/Boyett_et_al_2023.py
@@ -1,0 +1,40 @@
+dictionary_tag         = "Boyett et al. 2023 (subm.)"
+
+reference              = "Boyett, Trenti, Leethochawalit, Calabro, Metha, Roberts-Borsani"
+
+url                    = "https://ui.adsabs.harvard.edu/abs/2023arXiv230300306B/abstract"
+
+description            = \
+"""
+From high-resolution spatially-resolved NIRSpec spectroscopy of a z = 9.3127 bright galaxy.
+"""
+
+data_structure         = "points" #grid or points
+
+extracted              = False
+
+ndim                   = 2
+
+dimensions_descriptors = ["redshift", "log stellar mass [Msun]"]
+
+axes                   = [[9.3127, 9.40]]
+
+err_left               = [[0.0002, 0.11]]
+
+err_right              = [[0.0002, 0.10]]
+
+values                 = [7.6]
+
+err_up                 = [0.5]
+
+err_down               = [0.5]
+
+upper_lim              = False
+
+lower_lim              = False
+
+SFR                    = [25.0]
+
+err_left_SFR           = [4.0]
+
+err_right_SFR          = [3.0]
