+++ corecon/data/mass_gas_metallicity_relation/Nakajima_et_al_2023.py
@@ -1,0 +1,322 @@
+dictionary_tag         = "Nakajima etal. 2023 (subm.)"
+
+reference              = "Nakajima, Ouchi, Isobe, Harikane, Zhang, et al. 2023 (subm.)"
+
+url                    = "https://ui.adsabs.harvard.edu/abs/2023arXiv230112825N/abstract"
+
+description            = \
+"""
+From 135 galaxies at 4<z<10 identified in the ERO, GLASS, and CEERS programs on JWST.
+"""
+
+data_structure         = "points" #grid or points
+
+extracted              = False
+
+ndim                   = 2
+
+dimensions_descriptors = ["redshift", "log stellar mass [Msun]"]
+
+axes                   = [[8.496, 7.6], [6.378, 8.55], [7.665, 8.77], [5.275, 9.45], [7.66, 7.78], [7.881, 8.16], [7.874, 9.15], [7.877, 8.27],
+       [7.879, 8.38], [7.286, 8.51], [6.229, 8.23], [5.772, 9.09], [5.763, 9.22], [5.331, np.nan], [5.133, 8.87], [4.797, 7.42],
+       [4.584, 9.12], [4.02, 9.4], [4.015, 8.11], [3.951, 8.5], [8.005, 8.61], [8.873, 8.01], [7.104, 7.94], [6.212, 6.11],
+       [5.27, 8.72], [5.668, 7.23], [6.105, 7.83], [5.641, 8.41], [6.051, 7.34], [5.513, 9.03], [6.618, 7.6], [6.005, 8.45],
+       [5.766, 9.35], [7.029, 7.61], [7.179, 7.45], [7.179, 7.82], [5.664, 8.48], [7.115, 8.72], [7.064, 8.25], [5.668, 9.03],
+       [6.695, 7.46], [6.059, 9.1], [6.055, 8.06], [5.273, 8.76], [5.804, 7.66], [5.666, 9.03], [7.752, 8.44], [7.549, 8.7],
+       [7.471, 9.39], [4.899, 8.76], [6.961, 9.42], [6.934, 9.34], [5.626, 9.3], [6.257, 8.54], [7.167, 7.58], [4.243, 8.13],
+       [8.681, 10.12], [7.779, 8.76], [8.714, 8.54], [7.825, 8.12], [8.612, 8.8], [7.197, 8.38], [6.794, 8.63], [6.19, 8.91],
+       [6.994, 8.75], [6.3, np.nan], [6.957, 9.41], [6.93, 8.71], [8.179, 8.65], [6.566, 8.67], [7.451, 8.98], [4.997, 9.38],
+       [4.897, 8.81], [4.637, 9.61], [4.48, 8.91], [4.482, 8.62], [5.152, 8.45], [4.88, 8.56], [4.999, 8.89], [5.007, 8.53],
+       [5.503, 8.96], [5.504, 8.39], [4.307, 9.05], [4.999, 9.17], [4.548, 8.74], [4.283, 9.14], [4.492, 8.45], [5.375, 8.92],
+       [5.262, 8.91], [5.296, 9.11], [4.481, 9.15], [4.756, 9.24], [4.471, 8.8], [5.274, 8.87], [4.631, 8.58], [4.631, 8.85],
+       [6.108, 9.29], [4.588, 8.93], [5.036, 8.85], [4.888, 8.72], [4.192, 8.51], [6.204, 8.7], [4.792, 9.03], [4.631, 9.11],
+       [4.637, 8.81], [5.303, 8.58], [4.638, 8.55], [5.076, 9.9], [4.382, 8.85], [4.61, 8.56], [4.488, 9.79], [5.206, 9.54],
[…]
