+++ corecon/data/mass_gas_metallicity_relation/Williams_et_al_2023.py
@@ -1,0 +1,46 @@
+dictionary_tag         = "Williams et al. 2023"
+
+reference              = "Williams, Kelly, Wenlei, Brammer, Zitrin, Treu, Scarlata, et al.; Science 380, 416 (2023)"
+
+url                    = "https://ui.adsabs.harvard.edu/abs/2023Sci...380..416W/abstract"
+
+description            = \
+"""
+From NIRCam observations of a z=9.51 galaxy magnified by the RX J2129+0005 cluster.
+"""
+
+data_structure         = "points" #grid or points
+
+extracted              = False
+
+ndim                   = 2
+
+dimensions_descriptors = ["redshift", "log stellar mass [Msun]"]
+
+axes                   = [[9.51, 7.63]]
+
+err_left               = [[0.01, 0.24]]
+
+err_right              = [[0.01, 0.22]]
+
+values                 = [7.63]
+
+err_up                 = [0.07]
+
+err_down               = [0.05]
+
+upper_lim              = False
+
+lower_lim              = False
+
+MUV                    = [-17.4]
+
+err_left_MUV           = [0.22]
+
+err_right_MUV          = [0.22]
+
+SFR                    = [0.9]
+
+err_left_SFR           = [0.32]
+
+err_right_SFR          = [0.32]

[…]
