### 2023-07-05 1517af9
add ionizing_photons_production_efficiency field
