+++ corecon/data/ionizing_photons_production_efficiency/Simmonds_et_al_2023.py
@@ -1,0 +1,51 @@
+dictionary_tag         = "Simmonds et al. 2023"
+
+reference              = "Simmonds, Tacchella, Maseda, Williams, Baker, Witte, et al. subm."
+
+url                    = "https://ui.adsabs.harvard.edu/abs/20223MNRAS.tmp.1726S/abstract"
+
+description            = \
+"""
+From LAEs at 5.4 < z < 6.6 (robustly assigned from MUSE deep fields). LAEs are part of the JEMS survey from JWST.
+"""
+
+data_structure         = "points" #grid or points
+
+extracted              = False
+
+ndim                   = 1
+
+dimensions_descriptors = ["redshift"]
+
+axes                   = [5.443, 5.467, 5.468, 5.468, 5.479, 5.520, 5.520, 5.522, 5.822, 5.885, 5.889, 5.893, 5.895, 5.897, 5.899, 
+			  5.915, 5.915, 5.919, 5.930, 5.946, 5.972, 5.981, 6.096, 6.102, 6.106, 6.109, 6.302, 6.310, 6.342, 6.631 ]
+
+values                 = [25.61, 25.86, 25.66, 25.56, 26.59, 25.96, 25.26, 25.31, 24.48,
+			  25.44, 25.61, 25.44, 25.43, 25.57, 25.29, 26.03, 25.52, 25.72,
+			  25.22, 25.42, 25.16, 25.35, 25.15, 25.68, 25.53, 25.25, 26.21,
+			  25.32, 25.33, 24.92]
+
+err_up                 = [0.12, 0.13, 0.02, 0.05, 0.09, 0.06, 0.02, 0.02, 0.08, 0.03, 0.11,
+			  0.11, 0.07, 0.04, 0.11, 0.08, 0.1 , 0.11, 0.23, 0.1 , 0.11, 0.12,
+			  0.24, 0.01, 0.67, 0.3 , 0.1 , 0.05, 0.08, 0.06]
+
+err_down               = [0.11, 0.12, 0.02, 0.05, 0.1 , 0.07, 0.08, 0.09, 0.1 , 0.06, 0.13,
+			  0.14, 0.15, 0.04, 0.13, 0.11, 0.13, 0.1 , 0.3 , 0.07, 0.14, 0.09,
+			  0.33, 0.04, 1.  , 0.39, 0.09, 0.07, 0.11, 0.08]
+
+upper_lim              = False
+
+lower_lim              = False
+
+beta                   = [-1.6 , -1.84, -2.07, -1.95, -1.47, -1.75, -1.83, -2.24, -2.05,
+			  -2.26, -1.93, -2.43, -2.03, -2.35, -1.95, -2.66, -2.24, -1.78,
+			  -2.28, -2.22, -2.36, -2.42, -2.31, -2.16, -2.38, -2.46, -1.79,
+			  -2.01, -2.11, -1.82]
+
+err_up_beta            = [0.02, 0.01, 0.01, 0.01, 0.03, 0.03, 0.01, 0.02, 0.01, 0.02, 0.01,
+		          0.02, 0.01, 0.01, 0.01, 0.02, 0.02, 0.01, 0.02, 0.02, 0.01, 0.02,
+		          0.01, 0.01, 0.02, 0.02, 0.01, 0.01, 0.01, 0.01]
+
+err_down_beta          = [0.02, 0.01, 0.01, 0.01, 0.03, 0.03, 0.01, 0.02, 0.01, 0.02, 0.01,
+		          0.02, 0.01, 0.01, 0.01, 0.02, 0.02, 0.01, 0.02, 0.02, 0.01, 0.02,
+		          0.01, 0.01, 0.02, 0.02, 0.01, 0.01, 0.01, 0.01]
