+++ corecon/data/ionizing_photons_production_efficiency/Jung_et_al_2023.py
@@ -1,0 +1,34 @@
+dictionary_tag         = "Jung et al. 2023"
+
+reference              = "Jung, Finkelstein, Arrabal Haro, Dickinson, Ferguson, Hutchinson, et al., subm."
+
+url                    = "https://ui.adsabs.harvard.edu/abs/2023arXiv230405385J/abstract"
+
+description            = \
+"""
+From JWST/NIRSpec observations of three z~7.5 LAE detected with Keck/MOSFIRE. Part of CEERS. 
+"""
+
+data_structure         = "points" #grid or points
+
+extracted              = False
+
+ndim                   = 1
+
+dimensions_descriptors = ["redshift"]
+
+axes                   = [7.7528, 7.5457, 7.4710]
+
+err_left               = [0.0011, 0.0001, 0.0000]
+
+err_right              = [0.0011, 0.0001, 0.0000]
+
+values                 = [25.82, 25.24, 24.44]
+
+err_up                 = [0.39, 0.33, 0.16]
+
+err_down               = [0.39, 0.33, 0.16]
+
+upper_lim              = False
+
+lower_lim              = False
