+++ corecon/data/mass_gas_metallicity_relation/Fujimoto_et_al_2023.py
@@ -37,6 +37,6 @@ MUV                    = [-20.47, -20.75, -18.55 , -18.38, -19.08]
 
 SFR                    = [9.8, 3.9 , 64.3, 0.8 , 2.7]
 
-err_left_SFR           = [3.2, 0.6 , 50.8, 0.3 , 1.0 ]
+err_down_SFR           = [3.2, 0.6 , 50.8, 0.3 , 1.0 ]
 
-err_right_SFR          = [2.9, 1.1, 18.6, 0.7, 1.5]
+err_up_SFR             = [2.9, 1.1, 18.6, 0.7, 1.5]
