+++ corecon/data/ionizing_photons_production_efficiency/Endsley_et_al_2022.py
@@ -1,0 +1,40 @@
+dictionary_tag         = "Endsley et al. 2022"
+
+reference              = "Endsley, Stark, MNRAS 511, 6042 (2022)"
+
+url                    = "https://ui.adsabs.harvard.edu/abs/2022MNRAS.511.6042E/abstract"
+
+description            = \
+"""
+From 10 UV-bright z~6.7 spatially-close galaxies (in the vicinity of three LAEs) in the COSMOS field
+"""
+
+data_structure         = "points" #grid or points
+
+extracted              = False
+
+ndim                   = 1
+
+dimensions_descriptors = ["redshift"]
+
+axes                   = [6.882, 6.701, 6.702, 6.732, 6.73, 6.748, 6.814, 6.759, 6.487, 6.750, 6.67]
+
+err_left               = [0.0, 0.0, 0.0, 0.0, 0.04, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0]
+
+err_right              = [0.0, 0.0, 0.0, 0.0, 0.05, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0]
+
+values                 = [26.32, 25.72, 25.85, 25.68, 25.95, 25.88, 25.85, 25.76, 25.46, 25.75, 26.07]
+
+err_up                 = [0.18, 0.14, 0.21, 0.16, 0.16, 0.16, 0.16, 0.19, 0.24, 0.12, 0.18]
+
+err_down               = [0.18, 0.14, 0.19, 0.19, 0.23, 0.18, 0.12, 0.16, 0.26, 0.20, 0.20]
+
+upper_lim              = False
+
+lower_lim              = False
+
+M_UV                   = [-21.5, -21.1, -21.3, -21.0, -21.3, -20.4, -21.5, -20.6, -20.7, -22, -20.8]
+
+err_up_M_UV            = [0.1, 0.1, 0.1, 0.2, 0.1, 0.2, 0.1, 0.1, 0.2, 0.1, 0.1]
+
+err_down_M_UV          = [0.1, 0.1, 0.1, 0.1, 0.1, 0.2, 0.1, 0.1, 0.1, 0.1, 0.1]
