+++ corecon/version.py
@@ -1,1 +1,1 @@
-__version__ = '0.6.1'
+__version__ = '1.1'
