### 2023-07-05 ad6b498
filters now update derived variables
